+++ src/main/java/nextstep/subway/auth/ui/token/TokenAuthenticationInterceptor.java
@@ -7,8 +7,8 @@ import nextstep.subway.auth.domain.UserDetails;
 import nextstep.subway.auth.domain.UserDetailsService;
 import nextstep.subway.auth.dto.TokenResponse;
 import nextstep.subway.auth.infrastructure.JwtTokenProvider;
-import nextstep.subway.auth.ui.AuthenticationConverter;
-import nextstep.subway.auth.ui.AuthenticationInterceptor;
+import nextstep.subway.auth.ui.base.AuthenticationConverter;
+import nextstep.subway.auth.ui.base.AuthenticationInterceptor;
 import org.springframework.http.MediaType;
 
 import javax.servlet.http.HttpServletRequest;
